Rename some .c files to .cpp
