+++ libgadget/exchange.hpp
