+++ libgadget/sfr_eff.cpp
